+++ packages/amplify-e2e-tests/src/__tests__/auth_3.test.ts
@@ -1,25 +1,15 @@
-import * as fs from 'fs-extra';
+import { initJSProjectWithProfile, deleteProject, amplifyPushAuth } from 'amplify-e2e-core';
 import {
-  initJSProjectWithProfile,
-  initAndroidProjectWithProfile,
-  initIosProjectWithProfile,
-  deleteProject,
-  amplifyPushAuth,
-} from 'amplify-e2e-core';
-import {
-  addAuthWithRecaptchaTrigger,
-  addAuthWithCustomTrigger,
-  addAuthWithSignInSignOutUrl,
-  updateAuthWithoutCustomTrigger,
-  updateAuthRemoveRecaptchaTrigger,
-  updateAuthSignInSignOutUrl,
+  addAuthWithDefault,
+  removeAuthWithDefault,
+  addAuthWithMaxOptions,
+  addAuthUserPoolOnly,
+  getBackendAmplifyMeta,
 } from 'amplify-e2e-core';
 import {
   createNewProjectDir,
   deleteProjectDir,
   getProjectMeta,
-  getAwsAndroidConfig,
-  getAwsIOSConfig,
   getUserPool,
   getUserPoolClients,
   getLambdaFunction,
@@ -30,10 +20,10 @@ const defaultsSettings = {
   name: 'authTest',
 };
 
-describe('amplify updating auth...', () => {
+describe('amplify add auth...', () => {
   let projRoot: string;
   beforeEach(async () => {
-    projRoot = await createNewProjectDir('auth-update');
+    projRoot = await createNewProjectDir('auth');
   });
 
   afterEach(async () => {
@@ -41,74 +31,63 @@ describe('amplify updating auth...', () => {
     deleteProjectDir(projRoot);
   });
 
-  it('...should edit signin url on update', async () => {
-    const settings = {
-      signinUrl: 'http://localhost:3001/',
-      signoutUrl: 'http://localhost:3002/',
-      updatesigninUrl: 'http://localhost:3003/',
-      updatesignoutUrl: 'http://localhost:3004/',
-    };
-    await initAndroidProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithSignInSignOutUrl(projRoot, settings);
-    await updateAuthSignInSignOutUrl(projRoot, settings);
+  it('...should init a project and add auth with defaults and push, then remove auth and push should clean up trust relationship conditions', async () => {
+    await initJSProjectWithProfile(projRoot, defaultsSettings);
+    await addAuthWithDefault(projRoot, {});
+    await amplifyPushAuth(projRoot);
+
+    const amplifyMeta = getBackendAmplifyMeta(projRoot);
+    const { AuthRoleName, UnauthRoleName } = amplifyMeta.providers.awscloudformation;
+    const cognitoResource = Object.values(amplifyMeta.auth).find((res: any) => {
+      return res.service === 'Cognito';
+    }) as any;
+    const idpId = cognitoResource.output.IdentityPoolId;
+
+    expect(AuthRoleName).toHaveValidPolicyConditionMatchingIdpId(idpId);
+    expect(UnauthRoleName).toHaveValidPolicyConditionMatchingIdpId(idpId);
+
+    await removeAuthWithDefault(projRoot);
+    await amplifyPushAuth(projRoot);
+
+    expect(AuthRoleName).not.toHaveValidPolicyConditionMatchingIdpId(idpId);
+    expect(UnauthRoleName).not.toHaveValidPolicyConditionMatchingIdpId(idpId);
   });
 
-  it('...should init a project and add auth with a custom trigger, and then update to remove the custom js while leaving the other js', async () => {
+  it('...should init a project with only user pool and no identity pool', async () => {
     await initJSProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithCustomTrigger(projRoot, {});
+    await addAuthUserPoolOnly(projRoot, {});
+    await amplifyPushAuth(projRoot);
+    const meta = getProjectMeta(projRoot);
+    const id = Object.keys(meta.auth).map(key => meta.auth[key])[1].output.UserPoolId;
+    const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
+
+    expect(userPool.UserPool).toBeDefined();
+  });
+
+  it('...should init a project where all possible options are selected', async () => {
+    await initJSProjectWithProfile(projRoot, defaultsSettings);
+    await addAuthWithMaxOptions(projRoot, {});
     await amplifyPushAuth(projRoot);
     const meta = getProjectMeta(projRoot);
 
-    const functionName = `${Object.keys(meta.auth)[0]}PreSignup-integtest`;
+    const createFunctionName = `${Object.keys(meta.auth)[1]}CreateAuthChallenge-integtest`;
+    const defineFunctionName = `${Object.keys(meta.auth)[1]}DefineAuthChallenge-integtest`;
 
-    const authMeta = Object.keys(meta.auth).map(key => meta.auth[key])[0];
+    const authMeta = Object.keys(meta.auth).map(key => meta.auth[key])[1];
     const id = authMeta.output.UserPoolId;
     const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
     const clientIds = [authMeta.output.AppClientIDWeb, authMeta.output.AppClientID];
     const clients = await getUserPoolClients(id, clientIds, meta.providers.awscloudformation.Region);
 
-    const lambdaFunction = await getLambdaFunction(functionName, meta.providers.awscloudformation.Region);
-    const dirContents = fs.readdirSync(`${projRoot}/amplify/backend/function/${Object.keys(meta.auth)[0]}PreSignup/src`);
-    expect(dirContents.includes('custom.js')).toBeTruthy();
+    const createFunction = await getLambdaFunction(createFunctionName, meta.providers.awscloudformation.Region);
+    const defineFunction = await getLambdaFunction(defineFunctionName, meta.providers.awscloudformation.Region);
+
     expect(userPool.UserPool).toBeDefined();
     expect(clients).toHaveLength(2);
-    expect(lambdaFunction).toBeDefined();
-    expect(lambdaFunction.Configuration.Environment.Variables.MODULES).toEqual('email-filter-blacklist,custom');
+    expect(createFunction).toBeDefined();
+    expect(defineFunction).toBeDefined();
 
-    await updateAuthWithoutCustomTrigger(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    const updatedFunction = await getLambdaFunction(functionName, meta.providers.awscloudformation.Region);
-    const updatedDirContents = fs.readdirSync(`${projRoot}/amplify/backend/function/${Object.keys(meta.auth)[0]}PreSignup/src`);
-    expect(updatedDirContents.includes('custom.js')).toBeFalsy();
-    expect(updatedDirContents.includes('email-filter-blacklist.js')).toBeTruthy();
-    expect(updatedFunction.Configuration.Environment.Variables.MODULES).toEqual('email-filter-blacklist');
-  });
-
-  it('...should init an android project and add customAuth flag, and remove flag when custom auth triggers are removed upon update ', async () => {
-    await initAndroidProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithRecaptchaTrigger(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    let meta = getAwsAndroidConfig(projRoot);
-    expect(meta.Auth.Default.authenticationFlowType).toBeDefined();
-    expect(meta.Auth.Default.authenticationFlowType).toEqual('CUSTOM_AUTH');
-    await updateAuthRemoveRecaptchaTrigger(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    meta = getAwsAndroidConfig(projRoot);
-    expect(meta.Auth.Default.authenticationFlowType).toBeDefined();
-    expect(meta.Auth.Default.authenticationFlowType).toEqual('USER_SRP_AUTH');
-  });
-
-  it('...should init an ios project and add customAuth flag, and remove the flag when custom auth triggers are removed upon update', async () => {
-    await initIosProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithRecaptchaTrigger(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    let meta = getAwsIOSConfig(projRoot);
-    expect(meta.Auth.Default.authenticationFlowType).toBeDefined();
-    expect(meta.Auth.Default.authenticationFlowType).toEqual('CUSTOM_AUTH');
-    await updateAuthRemoveRecaptchaTrigger(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    meta = getAwsIOSConfig(projRoot);
-    expect(meta.Auth.Default.authenticationFlowType).toBeDefined();
-    expect(meta.Auth.Default.authenticationFlowType).toEqual('USER_SRP_AUTH');
+    expect(createFunction.Configuration.Environment.Variables.MODULES).toEqual('custom');
+    expect(defineFunction.Configuration.Environment.Variables.MODULES).toEqual('custom');
   });
 });
