+++ packages/amplify-e2e-tests/src/__tests__/schema-auth-11.test.ts
@@ -14,13 +14,18 @@ describe('api directives @auth batch 3', () => {
     deleteProjectDir(projectDir);
   });
 
-  it('auth staticGroup1', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'staticGroup1');
+  it('auth dynamicGroup1', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'dynamicGroup1');
     expect(testresult).toBeTruthy();
   });
 
-  it('auth staticGroup2', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'staticGroup2');
+  it('auth dynamicGroup2', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'dynamicGroup2');
+    expect(testresult).toBeTruthy();
+  });
+
+  it('auth dynamicGroup3', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'dynamicGroup3');
     expect(testresult).toBeTruthy();
   });
 });
